feat: memperbaiki gambar yang tidak muncul dan refactor kode
