+++ server/controllers/main.js
@@ -1,6 +1,6 @@
 const path = require("path");
 
-const markdownParser = require("../utils/markdownParser");
+const { markdownParser } = require("../utils");
 const { commonRegex, statics } = require("../config/constant");
 
 const commonFolder = require("./commonFolder");
@@ -9,10 +9,7 @@ function main(req, res) {
   const originalURL = path.normalize(req.originalUrl);
 
   if (originalURL === "/") {
-    const md = markdownParser("README.md").replace(
-      "CONTRIBUTING.md",
-      "https://github.com/bellshade/Javascript/blob/main/CONTRIBUTING.md"
-    );
+    const md = markdownParser("README.md", originalURL);
 
     res.render("root", { md, dirs: statics });
   } else if (commonRegex(originalURL)) {
