+++ server/utils/markdownParser.js
@@ -4,7 +4,9 @@ const commonmark = require("commonmark");
 
 const { ROOT } = require("../config/constant");
 
-const Parser = (readmePath) => {
+const imageSrcRegex = new RegExp('img src="(?!http(s?))', "g");
+
+const Parser = (readmePath, originalURL) => {
   const reader = new commonmark.Parser();
   const writer = new commonmark.HtmlRenderer();
 
@@ -18,7 +20,16 @@ const Parser = (readmePath) => {
   const markdown = fs.readFileSync(fullPath, "utf8");
 
   const parsed = reader.parse(markdown);
-  return writer.render(parsed);
+  const result = writer.render(parsed);
+
+  const replaceAll = result
+    .replace(
+      "CONTRIBUTING.md",
+      "https://github.com/bellshade/Javascript/blob/main/CONTRIBUTING.md"
+    )
+    .replace(imageSrcRegex, `img src="${path.join("/static", originalURL)}/`);
+
+  return replaceAll;
 };
 
 module.exports = Parser;
